backend/requirements.txt
Flask
Flask-Cors
gunicorn
werkzeug
google-generativeai